new dir structure & fix    - explain wouldn't explain **before** taking action
